+++ business/src/main/java/com/kevinguanchedarias/owgejava/entity/Unit.java
@@ -83,6 +83,9 @@ public class Unit extends CommonEntityWithImageStore<Integer>
 	@Column(name = "bypass_shield", nullable = false)
 	private Boolean bypassShield = false;
 
+	@Column(name = "is_invisible", nullable = false)
+	private Boolean isInvisible = false;
+
 	/**
 	 * @return the hasToDisplayInRequirements
 	 * @since 0.9.0
@@ -307,6 +310,24 @@ public class Unit extends CommonEntityWithImageStore<Integer>
 	 */
 	public void setBypassShield(Boolean bypassShield) {
 		this.bypassShield = bypassShield;
+	}
+
+	/**
+	 * @return the isInvisible
+	 * @since 0.10.0
+	 * @author Kevin Guanche Darias <kevin@kevinguanchedarias.com>
+	 */
+	public Boolean getIsInvisible() {
+		return isInvisible;
+	}
+
+	/**
+	 * @param isInvisible the isInvisible to set
+	 * @author Kevin Guanche Darias
+	 * @since 0.10.0
+	 */
+	public void setIsInvisible(Boolean isInvisible) {
+		this.isInvisible = isInvisible;
 	}
 
 }